refactor: 💡 Use bigserial; unify API type
Now model API user type `Maybe _` as return value, where `Nothing` means
failure (no resource found/no changes done)

+++ src/Model/Article/Type.hs
@@ -2,13 +2,13 @@
 
 module Model.Article.Type where
 
-import           Data.Int                   (Int64)
 import           Data.Profunctor.Product.TH (makeAdaptorAndInstance)
 import           Data.Text                  (Text)
 import           Model.F                    (F)
 import           Model.TH                   (genNewtypeT, makeTypeInstanceFWR)
 import           Model.User.Type
 import           Opaleye
+import Data.Int (Int64)
 
 type Title = Text
 type Content = Text
@@ -18,8 +18,8 @@ makeAdaptorAndInstance "pArticleID" ''ArticleIDT
 
 data ArticleT a b c d
   = Article          -- ^ @TABLE article_table@
-    { articleID :: a -- ^ @article_id serial PRIMARY KEY@
-    , authorID  :: b -- ^ @author_id int REFERENCES blogger_table(blogger_id)@
+    { articleID :: a -- ^ @article_id bigserial PRIMARY KEY@
+    , authorID  :: b -- ^ @author_id bigint REFERENCES blogger_table(blogger_id)@
     , title     :: c -- ^ @title text NOT NULL@
     , content   :: d -- ^ @content text NOT NULL@
     -- TODO: rating of the article
